* Small refactoring to make the creation of extra SceneScreens easier
