+++ blade-engine/src/com/bladecoder/engine/ui/DebugScreen.java
@@ -129,14 +129,13 @@ public class DebugScreen implements BladeScreen {
 
 		// ------------- RECORDING
 
-		Recorder r = ((SceneScreen) ui.getScreen(Screens.SCENE_SCREEN)).getRecorder();
+		final Recorder r = ui.getRecorder();
 		TextButton play = new TextButton(r.isPlaying() ? "Stop" : "Play", ui.getSkin());
 		rec = new TextButton(r.isRecording() ? "Stop Rec" : "Rec", ui.getSkin());
 		play.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				Recorder r = scnScr.getRecorder();
+				final Recorder r = ui.getRecorder();
 
 				if (!r.isPlaying()) {
 					r.setFilename(recordings.getSelected());
@@ -153,8 +152,7 @@ public class DebugScreen implements BladeScreen {
 		rec.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				Recorder r = scnScr.getRecorder();
+				final Recorder r = ui.getRecorder();
 
 				if (r.isPlaying()) {
 					r.setPlaying(false);
@@ -233,14 +231,13 @@ public class DebugScreen implements BladeScreen {
 		table.add(scGroup);
 
 		// ------------- TESTERBOT
-		TesterBot bot = ((SceneScreen) ui.getScreen(Screens.SCENE_SCREEN)).getTesterBot();
+		final TesterBot bot = ui.getTesterBot();
 		
 		TextButton runBot = new TextButton(bot.isEnabled()?"Stop":"Run", ui.getSkin());
 		runBot.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				TesterBot bot = scnScr.getTesterBot();
+				final TesterBot bot = ui.getTesterBot();
 				
 				bot.setMaxWaitInverval(Float.parseFloat(testerTimeConf.getText()));
 				bot.setInSceneTime(Float.parseFloat(inSceneTimeConf.getText()));
@@ -265,8 +262,7 @@ public class DebugScreen implements BladeScreen {
 		testerLeaveConf.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				TesterBot bot = scnScr.getTesterBot();
+				final TesterBot bot = ui.getTesterBot();
 				
 				bot.setRunLeaveVerbs(!bot.isRunLeaveVerbs());
 			}
@@ -278,8 +274,7 @@ public class DebugScreen implements BladeScreen {
 		testerGotoConf.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				TesterBot bot = scnScr.getTesterBot();
+				final TesterBot bot = ui.getTesterBot();
 				
 				bot.setRunGoto(!bot.isRunGoto());
 			}
@@ -291,8 +286,7 @@ public class DebugScreen implements BladeScreen {
 		testerPassText.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				TesterBot bot = scnScr.getTesterBot();
+				final TesterBot bot = ui.getTesterBot();
 				
 				bot.setPassTexts(!bot.isPassTexts());
 			}
@@ -304,8 +298,7 @@ public class DebugScreen implements BladeScreen {
 		testerWaitWhenWalking.addListener(new ClickListener() {
 
 			public void clicked(InputEvent event, float x, float y) {
-				SceneScreen scnScr = (SceneScreen) ui.getScreen(Screens.SCENE_SCREEN);
-				TesterBot bot = scnScr.getTesterBot();
+				final TesterBot bot = ui.getTesterBot();
 				
 				bot.setWaitWhenWalking(!bot.isWaitWhenWalking());
 			}
